+++ src/main/java/de/gematik/idp/tests/aforeport/AfoSerenityTestResultParser.java
@@ -19,65 +19,14 @@ package de.gematik.idp.tests.aforeport;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-@Getter
 @Slf4j
-public class AfoSerenityTestParser implements ITestParser, ITestResultParser {
-
-    private final Map<String, List<Testcase>> parsedTestcasesPerAfo = new HashMap<>();
-
-    @Override
-    public void resetMap() {
-        parsedTestcasesPerAfo.clear();
-    }
-
-    public void parseDirectory(final File rootDir) {
-        if (rootDir == null) {
-            log.warn("Invalid test source NULL root dir");
-        } else {
-            final File[] files = rootDir.listFiles();
-            if (files == null) {
-                if (log.isWarnEnabled()) {
-                    log.warn(String.format("Invalid test source root dir %s", rootDir.getAbsolutePath()));
-                }
-            } else {
-                Arrays.stream(files)
-                    .filter(f -> f.isFile() && f.getName().endsWith(".json"))
-                    .forEach(this::inspectFile);
-            }
-        }
-    }
-
-    private void inspectFile(final File f) {
-        try {
-            final String gherkin = Files.readString(f.toPath());
-            final JSONObject jso = new JSONObject(gherkin);
-            if (jso.has("tags")) {
-                jso.getJSONArray("tags").forEach(tag -> {
-                    final JSONObject jsoTag = (JSONObject) tag;
-                    if ("Afo".equals(jsoTag.getString("type"))) {
-                        final String afoid = jsoTag.getString("name");
-                        parsedTestcasesPerAfo.computeIfAbsent(afoid, k -> new ArrayList<>());
-                        final Testcase tc = new Testcase();
-                        setTestCaseClassNMethod(jso, tc);
-                        parsedTestcasesPerAfo.get(afoid).add(tc);
-                    }
-                });
-            }
-
-        } catch (final IOException ioe) {
-            log.error("Failed to parse BDD", ioe);
-        }
-    }
+public class AfoSerenityTestResultParser implements ITestResultParser {
 
     @Override
     public void parseDirectoryForResults(final Map<String, TestResult> results, final File rootDir) {
@@ -98,14 +47,10 @@ public class AfoSerenityTestParser implements ITestParser, ITestResultParser {
     }
 
     // TO DO move to ctor for Testcase with JSONObject as param
-    private void setTestCaseClassNMethod(final JSONObject jso, final Testcase tc) {
-        final String id = jso.getString("id");
-        final String[] idarr = id.split(";");
-        tc.setClazz(String.join(".", Arrays.copyOf(idarr, idarr.length - 1)));
-        tc.setMethod(idarr[idarr.length - 1]);
-    }
-
     private void inspectFileForResults(final File f, final Map<String, TestResult> results) {
+        if (f.getName().equals("requirements.json")) {
+            return;
+        }
         try {
             final String gherkin = Files.readString(f.toPath());
             final JSONObject jso = new JSONObject(gherkin);
@@ -128,8 +73,18 @@ public class AfoSerenityTestParser implements ITestParser, ITestResultParser {
             results.put(tr.getClazz() + ":" + tr.getMethod(), tr);
 
         } catch (final IOException | JSONException ioe) {
-            log.error("Failed to parse BDD", ioe);
+            log.error("Failed to parse BDD file " + f.getAbsolutePath(), ioe);
         }
+    }
+
+    private void setTestCaseClassNMethod(final JSONObject jso, final Testcase tc) {
+        final String id = jso.getString("id");
+        final String[] idarr = id.split(";");
+        tc.setClazz(String.join(".", Arrays.copyOf(idarr, idarr.length - 1)));
+        tc.setMethod(idarr[idarr.length - 1]);
+        tc.setFeatureName(jso.getJSONObject("userStory").getString("storyName"));
+        tc.setScenarioName(jso.getString("title"));
+        tc.setPath(jso.getJSONObject("userStory").getString("path"));
     }
 
     private Result mapSerenityStatus(final String serenityStatus) {