+++ src/main/java/de/gematik/idp/tests/aforeport/ITestParser.java
@@ -22,6 +22,4 @@ import java.util.Map;
 public interface ITestParser {
 
     Map<String, List<Testcase>> getParsedTestcasesPerAfo();
-
-    void resetMap();
 }