+++ src/main/java/de/gematik/idp/tests/aforeport/Testcase.java
@@ -16,6 +16,7 @@
 
 package de.gematik.idp.tests.aforeport;
 
+import java.util.Objects;
 import lombok.Data;
 
 @Data
@@ -23,4 +24,21 @@ public class Testcase {
 
     String clazz;
     String method;
+    String featureName;
+    String scenarioName;
+    String path;
+
+    @Override
+    public boolean equals(final Object o) {
+        if (o instanceof Testcase) {
+            return Objects.equals(((Testcase) o).clazz, clazz) && Objects.equals(((Testcase) o).method,
+                method);
+        }
+        return false;
+    }
+
+    @Override
+    public int hashCode() {
+        return (clazz + ":" + method).hashCode();
+    }
 }