+++ src/test/java/de/gematik/idp/tests/aforeport/TestAfoSerenityTestParser.java
@@ -3,7 +3,7 @@ package de.gematik.idp.tests.aforeport;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.nio.file.Paths;
-import java.util.List;
+import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
@@ -11,24 +11,23 @@ public class TestAfoSerenityTestParser {
 
     @Test
     public void parseSerenityResultsOK() {
-        final AfoSerenityTestParser parser = new AfoSerenityTestParser();
+        final AfoSerenityTestResultParser parser = new AfoSerenityTestResultParser();
 
-        parser.parseDirectory(Paths.get("src", "test", "resources", "bdd").toFile());
-        final Map<String, List<Testcase>> tcs = parser.getParsedTestcasesPerAfo();
+        final Map<String, TestResult> results = new HashMap<>();
+        parser.parseDirectoryForResults(results, Paths.get("src", "test", "resources", "bdd").toFile());
 
-        assertThat(tcs).containsOnlyKeys("A_20315-01", "A_20314");
-        assertThat(tcs.get("A_20315-01")).hasSize(2);
-        assertThat(tcs.get("A_20315-01").get(0).getMethod())
-            .isEqualTo("gettoken-mit-sso-token---veralteter-sso-token-wird-abgelehnt");
+        assertThat(results).containsOnlyKeys(
+            "fordere-access-token-mittels-sso-token-an:gettoken-mit-sso-token---veralteter-sso-token-wird-abgelehnt",
+            "fordere-access-token-mit-einer-signierten-challenge-an:gettoken-signierte-challenge---veralteter-token-code-wird-abgelehnt");
     }
 
     @Test
     public void parseSerenityResultsInvalidRoot() {
-        final AfoSerenityTestParser parser = new AfoSerenityTestParser();
+        final AfoSerenityTestResultParser parser = new AfoSerenityTestResultParser();
 
-        parser.parseDirectory(Paths.get("src", "test", "resources", "bdd-NonExisting").toFile());
-        final Map<String, List<Testcase>> tcs = parser.getParsedTestcasesPerAfo();
+        final Map<String, TestResult> results = new HashMap<>();
+        parser.parseDirectoryForResults(results, Paths.get("src", "test", "resources", "bdd-NonExisting").toFile());
 
-        assertThat(tcs.keySet()).hasSize(0);
+        assertThat(results.keySet()).hasSize(0);
     }
 }